+++ pkgs/code_builder/test/specs/function_type_test.dart
@@ -50,10 +50,9 @@ void main() {
 
     test('should support toTypeDef', () {
       final funcType = FunctionType(
-        (b) =>
-            b
-              ..returnType = refer('String')
-              ..requiredParameters.add(refer('int')),
+        (b) => b
+          ..returnType = refer('String')
+          ..requiredParameters.add(refer('int')),
       );
       expect(
         funcType.toTypeDef('MyFunc'),